edit ss init

+++ service/shadowsocks-libev-debian.sh
@@ -24,6 +24,17 @@ RET_VAL=0
 
 [ -x $DAEMON ] || exit 0
 
+check_pid(){
+	get_pid=`ps -ef |grep -v grep | grep ss-server |awk '{print $2}'`
+}
+
+check_pid
+if [ -z $get_pid ]; then
+    if [ -e $PID_FILE ]; then
+        rm -f $PID_FILE
+    fi
+fi
+
 if [ ! -d $PID_DIR ]; then
     mkdir -p $PID_DIR
     if [ $? -ne 0 ]; then
@@ -38,13 +49,15 @@ if [ ! -f $CONF ]; then
 fi
 
 check_running() {
-    if [ -r $PID_FILE ]; then
-        read PID < $PID_FILE
-        if [ -d "/proc/$PID" ]; then
-            return 0
-        else
-            rm -f $PID_FILE
-            return 1
+    if [ -e $PID_FILE ]; then
+        if [ -r $PID_FILE ]; then
+            read PID < $PID_FILE
+            if [ -d "/proc/$PID" ]; then
+                return 0
+            else
+                rm -f $PID_FILE
+                return 1
+            fi
         fi
     else
         return 2
@@ -69,7 +82,9 @@ do_start() {
         echo "$NAME (pid $PID) is already running..."
         return 0
     fi
-    $DAEMON -v -c $CONF -f $PID_FILE
+    $DAEMON -c $CONF > /dev/null 2>&1 &
+    check_pid
+    echo $get_pid > $PID_FILE
     if check_running; then
         echo "Starting $NAME success"
     else
